fix merge conflicts from courses taken to course view
