--- client/src/app/course.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { CourseService } from './course.service';
-
-describe('CourseService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: CourseService = TestBed.get(CourseService);
-    expect(service).toBeTruthy();
-  });
-});
